Update the main python script to python3
